Separate delegators into separate files
